pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
]
